Adjust PR comments: - remove payBlockReward from BlockExecution and use blockPreparator.payBlockReward instead - remove executeTransaction from BlockExecution and use blockPreparator.executeTransaction instead - move test connected with blockPreparator to separate class - remove .get call in importBlockToTop - simplyfy resolveBranch method
